chatbotbackend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-community==0.0.19
pydantic==2.6.0
websockets==12.0
python-dotenv==1.0.0
